Move ParticleDataFlags to hoomd/custom_action.py
