+++ hoomd/__init__.py
@@ -75,7 +75,7 @@ from hoomd.snapshot import Snapshot
 from hoomd.logger import Logger
 from hoomd.box import Box
 from hoomd import tuner
-from hoomd.custom_action import CustomAction
+from hoomd.custom_action import CustomAction, ParticleDataFlags
 from hoomd.custom_operation import _CustomOperation
 
 from hoomd import logger
