+++ hoomd/util.py
@@ -9,7 +9,6 @@ R""" Utilities.
 from numpy import ndarray
 from collections.abc import Iterable, Mapping
 from copy import deepcopy
-from enum import IntEnum
 
 
 ## \internal
@@ -249,9 +248,3 @@ def array_to_strings(value):
         return string_list
     else:
         return value
-
-
-class ParticleDataFlags(IntEnum):
-    PRESSURE_TENSOR = 0
-    ROTATIONAL_KINETIC_ENERGY = 1
-    EXTERNAL_FIELD_VIRIAL = 2
